requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
pandas==2.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
soupsieve==2.5
tzdata==2024.1
urllib3==2.2.1
gspread==6.1.4
google-auth==2.37.0
